+++ hpc/metax/smi.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+ht-smi topo --matrix