+++ coredump/info.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+sysctl kernel.core_pipe_limit