+++ security/openssh/crypt.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# /etc/crypto-policies/back-ends/openssh.config
+# /etc/crypto-policies/back-ends/opensshserver.config
+set -e
+
+# $ scp -c [cipher] ...
+ssh -Q cipher