+++ fs/procfs/kallsyms/ksym_helpers.c
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+/* Copyright (c) 2025 Rong Tao */
+