+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 #### Changed
 #### Deprecated
 #### Removed
+- remove tlconfig;
 #### Fixed
 #### Security
 #### Docs